pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "saas-graph"
version = "0.1.0"
description = "AI analytics assistant framework — connect your database, get natural language querying"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.9"
authors = [{ name = "Alisia Reveli" }]
keywords = ["ai", "analytics", "text-to-sql", "rag", "llm", "saas", "multi-tenant"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Database",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "pydantic>=2.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
openai = ["openai>=1.0", "instructor>=1.0"]
postgres = ["asyncpg>=0.29"]
supabase = ["supabase>=2.0"]
pgvector = ["openai>=1.0"]
redis = ["redis>=5.0"]
tavily = ["tavily-python>=0.4"]
server = ["fastapi>=0.110", "uvicorn>=0.27", "sse-starlette>=2.0"]
langgraph = ["langgraph>=0.2"]
cli = ["typer>=0.12", "rich>=13.0"]
all = [
    "saas-graph[openai,postgres,pgvector,redis,tavily,server,langgraph,cli]",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.4",
]

[project.urls]
Homepage = "https://github.com/alisia-reveli/saas-graph"
Documentation = "https://github.com/alisia-reveli/saas-graph#readme"
Repository = "https://github.com/alisia-reveli/saas-graph"
Issues = "https://github.com/alisia-reveli/saas-graph/issues"

[project.scripts]
saas-graph = "saas_graph.cli.main:app"

[tool.hatch.build.targets.wheel]
packages = ["src/saas_graph"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
